+++ src/plugins/auto_batch/src/auto_batch.cpp
@@ -3,8 +3,6 @@
 //
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
-#include "auto_batch.hpp"
-
 #include <iostream>
 #include <map>
 #include <memory>
@@ -13,15 +11,21 @@
 #include <utility>
 #include <vector>
 
+#include "async_infer_request.hpp"
+#include "compiled_model.hpp"
 #include "cpp_interfaces/interface/ie_internal_plugin_config.hpp"
 #include "dimension_tracker.hpp"
 #include "ie_icore.hpp"
+#include "ie_metric_helpers.hpp"
 #include "ie_ngraph_utils.hpp"
 #include "ie_performance_hints.hpp"
 #include "openvino/pass/manager.hpp"
 #include "openvino/runtime/device_id_parser.hpp"
 #include "openvino/runtime/intel_gpu/properties.hpp"
 #include "openvino/util/common_util.hpp"
+#include "plugin.hpp"
+#include "sync_infer_request.hpp"
+#include "threading/ie_thread_safe_containers.hpp"
 #include "transformations/common_optimizations/dimension_tracking.hpp"
 #include "transformations/init_node_info.hpp"
 #include "transformations/utils/utils.hpp"
